requirements.txt
asgiref==3.3.1
attrs==20.3.0
coverage==5.4
Django==3.1.5
django-debug-toolbar==3.2
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==5.8.0
iniconfig==1.1.1
packaging==20.8
path==15.1.0
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
